Move DrawView to separate file
